refactor(runner): move the URL helpers next to the URL code
ServerHostFromURL and sanitizeURLError sat in tunnel_client.go between the
constructor and the registry, even though the package already keeps its URL
helpers in ws_url.go and the host helper now serves the FTP handler too. Both
move there with their constant, and their tests move to ws_url_test.go. What is
left in tunnel_client.go is the client itself plus closeWriteSide, which sits
right after its only caller.

The shared dial constants in tunnel_daemon.go move up to the top of the file,
matching where tunnel_relay_unix.go declares its own.

No behavior change: symbols only moved, and the now-unused errors and net/url
imports are dropped from tunnel_client.go.
